Resize dialog windows which exceeds the screen width/height
Firefox for instance spawns lots of dialogs which are bigger than the
actual screen (e.g. the "save page under"-dialog). This commit
introduces a check for ensuring that all new dialog windows fit on the
screen.

Since it is also somewhat annoying if a dialog window fits into the
screen but takes up the entire screen this commit introduces a tolerance
in percent.